+++ vscode/test/e2e/attribution.test.ts
@@ -1,8 +1,8 @@
-import { type Frame, type FrameLocator, type Locator, type Page, expect } from '@playwright/test'
+import { expect } from '@playwright/test'
 
 import * as mockServer from '../fixtures/mock-server'
 
-import { sidebarSignin } from './common'
+import { createEmptyChatPanel, sidebarSignin } from './common'
 import {
     type DotcomUrlOverride,
     type ExpectedEvents,
@@ -35,7 +35,8 @@ const test = baseTest
 
 test('attribution search enabled in chat', async ({ page, sidebar, expectedEvents }) => {
     await fetch(`${mockServer.SERVER_URL}/.test/attribution/enable`, { method: 'POST' })
-    const [chatFrame, chatInput] = await prepareChat2(page, sidebar)
+    await sidebarSignin(page, sidebar)
+    const [chatFrame, chatInput] = await createEmptyChatPanel(page)
     await chatInput.fill('show me a code snippet')
     await chatInput.press('Enter')
     await expect(chatFrame.getByTestId('attribution-indicator')).toBeVisible()
@@ -43,17 +44,9 @@ test('attribution search enabled in chat', async ({ page, sidebar, expectedEvent
 
 test('attribution search disabled in chat', async ({ page, sidebar, expectedEvents }) => {
     await fetch(`${mockServer.SERVER_URL}/.test/attribution/disable`, { method: 'POST' })
-    const [chatFrame, chatInput] = await prepareChat2(page, sidebar)
+    await sidebarSignin(page, sidebar)
+    const [chatFrame, chatInput] = await createEmptyChatPanel(page)
     await chatInput.fill('show me a code snippet')
     await chatInput.press('Enter')
     await expect(chatFrame.getByTestId('attribution-indicator')).toBeHidden()
 })
-
-async function prepareChat2(page: Page, sidebar: Frame): Promise<[FrameLocator, Locator]> {
-    await sidebarSignin(page, sidebar)
-    await page.getByRole('button', { name: 'New Chat', exact: true }).click()
-    // Chat webview iframe is the second and last frame (search is the first)
-    const chatFrame = page.frameLocator('iframe.webview').last().frameLocator('iframe')
-    const chatInput = chatFrame.getByRole('textbox', { name: 'Chat message' })
-    return [chatFrame, chatInput]
-}
