+++ vscode/test/e2e/chat-edits.test.ts
@@ -1,7 +1,7 @@
 import { expect } from '@playwright/test'
 
 import { isMacOS } from '@sourcegraph/cody-shared'
-import { sidebarSignin } from './common'
+import { createEmptyChatPanel, sidebarSignin } from './common'
 import { type ExpectedEvents, test, withPlatformSlashes } from './helpers'
 
 const osKey = isMacOS() ? 'Meta' : 'Control'
@@ -21,10 +21,7 @@ test.extend<ExpectedEvents>({
 })('editing follow-up messages in chat view', async ({ page, sidebar }) => {
     await sidebarSignin(page, sidebar)
 
-    await page.getByRole('button', { name: 'New Chat', exact: true }).click()
-
-    const chatFrame = page.frameLocator('iframe.webview').last().frameLocator('iframe')
-    const chatInput = chatFrame.getByRole('textbox', { name: 'Chat message' })
+    const [chatFrame, chatInput] = await createEmptyChatPanel(page)
 
     // Chat Action Buttons - above the input box
     const editLastMessageButton = chatFrame.getByRole('button', { name: /^Edit Last Message / })
