+++ vscode/test/e2e/chat-history.test.ts
@@ -1,6 +1,6 @@
 import { expect } from '@playwright/test'
 
-import { sidebarSignin } from './common'
+import { createEmptyChatPanel, sidebarSignin } from './common'
 import { type ExpectedEvents, test } from './helpers'
 
 test.extend<ExpectedEvents>({
@@ -25,11 +25,8 @@ test.extend<ExpectedEvents>({
     const heyTreeItem = page.getByRole('treeitem', { name: 'Hey' })
     const holaTreeItem = page.getByRole('treeitem', { name: 'Hola' })
 
-    await page.getByRole('button', { name: 'New Chat', exact: true }).click()
+    const [chatPanelFrame, chatInput] = await createEmptyChatPanel(page)
 
-    const chatPanelFrame = page.frameLocator('iframe.webview').last().frameLocator('iframe')
-
-    const chatInput = chatPanelFrame.getByRole('textbox', { name: 'Chat message' })
     await chatInput.fill('Hey')
     await chatInput.press('Enter')
