+++ vscode/test/e2e/chat-rateLimit.test.ts
@@ -1,8 +1,8 @@
-import { type Frame, type FrameLocator, type Locator, type Page, expect } from '@playwright/test'
+import { expect } from '@playwright/test'
 
 import * as mockServer from '../fixtures/mock-server'
 
-import { sidebarSignin } from './common'
+import { createEmptyChatPanel, sidebarSignin } from './common'
 import { type DotcomUrlOverride, type ExpectedEvents, test as baseTest } from './helpers'
 
 const test = baseTest.extend<DotcomUrlOverride>({ dotcomUrl: mockServer.SERVER_URL })
@@ -25,7 +25,8 @@ test.extend<ExpectedEvents>({
         method: 'POST',
     })
 
-    const [chatFrame, chatInput] = await prepareChat(page, sidebar)
+    await sidebarSignin(page, sidebar)
+    const [chatFrame, chatInput] = await createEmptyChatPanel(page)
     await chatInput.fill('test message')
     await chatInput.press('Enter')
 
@@ -50,7 +51,8 @@ test.extend<ExpectedEvents>({
         method: 'POST',
     })
 
-    const [chatFrame, chatInput] = await prepareChat(page, sidebar)
+    await sidebarSignin(page, sidebar)
+    const [chatFrame, chatInput] = await createEmptyChatPanel(page)
     await chatInput.fill('test message')
     await chatInput.press('Enter')
 
@@ -75,19 +77,11 @@ test.extend<ExpectedEvents>({
         method: 'POST',
     })
 
-    const [chatFrame, chatInput] = await prepareChat(page, sidebar)
+    await sidebarSignin(page, sidebar)
+    const [chatFrame, chatInput] = await createEmptyChatPanel(page)
     await chatInput.fill('test message')
     await chatInput.press('Enter')
 
     await expect(chatFrame.getByRole('heading', { name: 'Unable to Send Message' })).toBeVisible()
     await expect(chatFrame.getByRole('button', { name: 'Learn More' })).toBeVisible()
 })
-
-export async function prepareChat(page: Page, sidebar: Frame): Promise<[FrameLocator, Locator]> {
-    await sidebarSignin(page, sidebar)
-    await page.getByRole('button', { name: 'New Chat', exact: true }).click()
-    // Chat webview iframe is the second and last frame (search is the first)
-    const chatFrame = page.frameLocator('iframe.webview').last().frameLocator('iframe')
-    const chatInput = chatFrame.getByRole('textbox', { name: 'Chat message' })
-    return [chatFrame, chatInput]
-}
