+++ vscode/test/e2e/cody-ignore.test.ts
@@ -1,6 +1,6 @@
 import path from 'path'
 import { expect } from '@playwright/test'
-import { sidebarExplorer, sidebarSignin } from './common'
+import { createEmptyChatPanel, sidebarExplorer, sidebarSignin } from './common'
 import { type ExpectedEvents, test } from './helpers'
 
 /**
@@ -46,11 +46,9 @@ test.extend<ExpectedEvents>({
     await page.click('.badge[aria-label="Cody"]')
 
     // Start new chat
-    await page.getByRole('button', { name: 'New Chat', exact: true }).click()
+    const [chatPanel, chatInput] = await createEmptyChatPanel(page)
 
     /* TEST: Chat Context - Ignored file do not show up with context */
-    const chatPanel = page.frameLocator('iframe.webview').last().frameLocator('iframe')
-    const chatInput = chatPanel.getByRole('textbox', { name: 'Chat message' })
     await chatInput.focus()
     await chatInput.fill('Ignore me')
     await chatInput.press('Enter')
